📏 Some improvements
